+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Change Log
 ==========
 
+ * added java 1.8 as a run-time requirement
+ * changed format of 'path' argument to 'file' and 'files' functions
  * renamed 'concatenateStrings' to 'concatenateStringArrays'
  * renamed 'concatenateFiles' to 'concatenateFileArrays'
  * renamed 'concatenateBlobs' to 'concatenateBlobArrays'